+++ python/sglang/srt/model_executor/model_runner.py
@@ -155,7 +155,7 @@ from sglang.srt.utils.offloader import (
     set_offloader,
 )
 from sglang.srt.utils.patch_torch import monkey_patch_torch_reductions
-from sglang.srt.utils.pyt_hooks import PytHooks
+from sglang.srt.utils.pytorch_hooks import PytHooks
 from sglang.srt.utils.torch_memory_saver_adapter import TorchMemorySaverAdapter
 from sglang.srt.weight_sync.tensor_bucket import (
     FlattenedTensorBucket,
